+++ src/utils/dataset.js
@@ -1,9 +1,10 @@
 import {getHeader} from "./common";
 import config from "../app.config";
 
+
 export async function getDatasetsRequest(title, limit) {
 	// Clowder API to get dataset list
-	let url = `${config.hostname}/clowder/api/datasets?superAdmin=true&limit=${limit}`;
+	let url = `${config.hostname}/clowder/api/datasets&limit=${limit}`;
 	if (title) url = `${url}&title=${title}`;
 	const response = await fetch(url, {mode: "cors", headers: getHeader()});
 	if (response.status === 200) {
@@ -23,15 +24,13 @@ export async function getDatasetsRequest(title, limit) {
 }
 
 
-export async function createEmptyDatasetRequest(file) {
+export async function createEmptyDatasetRequest(dataset_name, dataset_description) {
 	// Clowder API call to create empty dataset
-	const url = `${config.hostname}/clowder/api/datasets/createempty?superAdmin=true`;
+	const url = `${config.hostname}/clowder/api/datasets/createempty`;
 	let authHeader = getHeader();
 	authHeader.append('Accept', 'application/json');
 	authHeader.append('Content-Type', 'application/json');
-	const datasetname = file.name.replace(/\.[^/.]+$/, ""); // get filename without extension as dataset name
-	const description = file.type;
-	const body_data = {"name": datasetname, "description": description};
+	const body_data = {"name": dataset_name, "description": dataset_description, "space": [config.space]};
 	const body = JSON.stringify(body_data);
 	const response = await fetch(url, {method:"POST", mode:"cors", headers:authHeader, body:body});
 	if (response.status === 200) {
@@ -50,9 +49,10 @@ export async function createEmptyDatasetRequest(file) {
 	}
 }
 
+
 export async function uploadFileToDatasetRequest(dataset_id, file) {
 	// Clowder API call to upload file to dataset
-	const upload_to_dataset_url = `${config.hostname}/clowder/api/uploadToDataset/${dataset_id}?extract=false`;
+	const upload_to_dataset_url = `${config.hostname}/clowder/api/uploadToDataset/${dataset_id}?extract=${config.extract}`;
 	let body = new FormData();
 	body.append("File" ,file);
 	let authHeader = getHeader();
@@ -76,22 +76,43 @@ export async function uploadFileToDatasetRequest(dataset_id, file) {
 	}
 }
 
-export async function checkHtmlInDatasetRequest(dataset_id){
-	// function to check if html file is there in the dataset
+
+export async function listFilesInDatasetRequest(dataset_id) {
+	// function to get a list of all files in clowder dataset
 	const listFiles_url = `${config.hostname}/clowder/api/datasets/${dataset_id}/listFiles`;
 	// get the list of files in dataset
-	const dataset_listFiles_response = await fetch(listFiles_url, {method:"GET", headers:getHeader(), mode: "cors"});
-	const dataset_listFiles = await dataset_listFiles_response.json();
-	// filter html file and select the first item in filtered array.
-	const htmlFile = Object.values(dataset_listFiles).filter(file => file.contentType === "text/html")[0];
-	// [ {"id":string, "size":string, "date-created":string, "contentType":text/html, "filename":string} ]
-	if (htmlFile !== undefined && htmlFile.contentType === "text/html") {
-		// found html file in dataset. return the object
-		console.log("html file generated");
-		return htmlFile;
+	const listFiles_response = await fetch(listFiles_url, {method:"GET", headers:getHeader(), mode: "cors"});
+	return listFiles_response.json();
+}
+
+
+export async function getFileInDataset(dataset_id, file_type, file_name=null){
+	// function to check if a specific file is present in dataset and return the file
+	// filter files on file type and filename and select the first item in filtered array.
+	let fileObjects = await listFilesInDatasetRequest(dataset_id);
+	console.log("getFileInDataset", dataset_id, file_type, file_name);
+	console.log("fileObjects", fileObjects);
+	let files = [];
+	if (file_name) {
+		files = Object.values(fileObjects).filter(file => {
+			if (file.contentType === file_type && file.filename === file_name){
+				return file
+			}
+		});
 	}
 	else {
-		console.log("html file generation failed");
+		files = Object.values(fileObjects).filter(file => file.contentType === file_type);
+	}
+	// get the first item of files list
+	const file = files[0];
+	// [ {"id":string, "size":string, "date-created":string, "contentType":text/html, "filename":string} ]
+	if (file !== undefined && file.contentType === file_type) {
+		// found file in dataset. return the object
+		console.log("File with type ", file_type, " generated");
+		return file;
+	}
+	else {
+		console.log("No File with type ", file_type);
 		return null;
 	}
 
@@ -105,7 +126,7 @@ export async function downloadDataset(datasetId, filename = null) {
 	} else {
 		filename = `${datasetId}.zip`;
 	}
-	let endpoint = `${config.hostname}/clowder/api/datasets/${datasetId}/download?superAdmin=true`;
+	let endpoint = `${config.hostname}/clowder/api/datasets/${datasetId}/download`;
 	let response = await fetch(endpoint, {method: "GET", mode: "cors", headers: await getHeader()});
 
 	if (response.status === 200) {